added dialog box
