+++ src/index.html
@@ -20,7 +20,7 @@
                 <section class="eventTaskContainer">
                     <section class="eventsContainer"></section>
                     <section class="tasksContainer">
-                        <section class="saveTaskButtonContainers"></section>
+                        <section class="addTaskButtonContainers"></section>
                         <section class="userTasksContainer"></section>
                     </section>
                 </section>
